refactor hex counts u16 -> i16

+++ client/src/map/draw.rs
@@ -11,12 +11,12 @@ pub fn draw_map(map_origin: &MapCoord) {
     let screen_height: i32 = unsafe { GetScreenHeight() };
     let origin_hex_coord: HexCoord = map_origin.hex_coord();
     let min_hex_coord: HexCoord = HexCoord {
-        i: if i32::from(origin_hex_coord.i) - 1 < 0 {
+        i: if origin_hex_coord.i - 1 < 0 {
             HEX_COUNT_SQRT - 1
         } else {
             origin_hex_coord.i - 1
         },
-        j: if i32::from(origin_hex_coord.j) - 1 < 0 {
+        j: if origin_hex_coord.j - 1 < 0 {
             HEX_COUNT_SQRT - 1
         } else {
             origin_hex_coord.j - 1
